+++ src/test/java/org/ow2/petals/deployer/utils/RuntimeModelComparatorTest.java
@@ -1,0 +1,18 @@
+package org.ow2.petals.deployer.utils;
+
+import java.io.IOException;
+import java.net.MalformedURLException;
+import java.net.URISyntaxException;
+
+import org.junit.Test;
+import org.ow2.petals.deployer.model.xml._1.Model;
+
+public class RuntimeModelComparatorTest {
+
+    @Test
+    public void equivalentModels() throws MalformedURLException, IOException, URISyntaxException {
+        Model model = ModelUtils.generateTestModel();
+
+    }
+
+}
